fix: page not found on touch devices, base layout, responsive
