+++ src/app/not-found.tsx
@@ -6,14 +6,19 @@ import { Link } from '@/i18n/routing'
 
 import '@/styles/globals.css'
 
+import { Metadata } from 'next'
 import { IconArrowLeft } from '@tabler/icons-react'
-import { NextIntlClientProvider } from 'next-intl'
-import { getMessages, getTranslations } from 'next-intl/server'
-import { ThemeProvider } from 'next-themes'
+import { getTranslations } from 'next-intl/server'
+
+export async function generateMetadata(): Promise<Metadata> {
+	const __ = await getTranslations('NotFound')
+
+	return {
+		title: __('title'),
+	}
+}
 
 export default async function NotFound() {
-	const messages = await getMessages()
-
 	const gifs = [
 		'https://media1.tenor.com/m/kSiC-0wGr4kAAAAC/monkey-technology.gif',
 		'https://media1.tenor.com/m/M0na3YR-rTcAAAAd/enter-dev.gif',
@@ -34,47 +39,55 @@ export default async function NotFound() {
 
 	const randomGif = gifs[Math.floor(Math.random() * gifs.length)]
 
-	const __ = await getTranslations('Default')
+	const __ = await getTranslations('NotFound')
 
 	return (
-		<html
-			translate="no"
-			className="text-content-light dark:text-content-dark layout:overflow-hidden w-full scroll-smooth bg-white font-sans antialiased dark:bg-black"
-			suppressHydrationWarning
-		>
-			<body className="text-content-light dark:text-content-dark layout:overflow-y-auto layout:h-dvh w-full overflow-x-hidden scroll-smooth">
-				<NextIntlClientProvider messages={messages}>
-					<ThemeProvider attribute="data-mode">
-						<FollowCursor
-							backgroundImage={randomGif}
-							rotationFactor={50}
-							cardWidth="400px"
-							offsetX={-200}
-							offsetY={-200}
-							className="pointer-events-none fixed inset-1/2 z-[40] opacity-0 transition-opacity duration-500 in-hover:!opacity-100"
-						/>
+		<>
+			<FollowCursor
+				backgroundImage={randomGif}
+				rotationFactor={50}
+				cardWidth="400px"
+				offsetX={-200}
+				offsetY={-200}
+				className="pointer-events-none fixed inset-1/2 z-[40] opacity-0 transition-opacity duration-500 in-hover:!opacity-100"
+			/>
 
-						<div className="flex min-h-dvh w-full flex-col items-center justify-center gap-6">
-							<h2 className="font-heading pr-4 text-[50vw] leading-none tracking-tighter sm:text-[40vw] md:text-[33vw] lg:text-[25vw] dark:text-white [&>span>span>span:last-child]:!hidden">
-								<SplitText text="404" delay={200} />
-							</h2>
+			<div className="only-touch:opacity-100 absolute inset-0 bg-black opacity-0" />
 
-							<AnimatedContent className="relative z-50">
-								<Button
-									as={Link}
-									href="/"
-									className="max-xs:w-[90vw] !h-14 font-medium backdrop-blur-sm md:!h-16"
-									size="lg"
-									variant="outline"
-								>
-									<IconArrowLeft />
-									{__('404')}
-								</Button>
-							</AnimatedContent>
-						</div>
-					</ThemeProvider>
-				</NextIntlClientProvider>
-			</body>
-		</html>
+			<div
+				className="only-touch:opacity-25 absolute inset-0 bg-cover bg-center opacity-0"
+				style={{ backgroundImage: `url(${randomGif})` }}
+			/>
+
+			<div className="flex min-h-dvh w-full flex-col items-center justify-center gap-6">
+				<h2 className="font-heading only-touch:text-white pr-4 text-[50vw] leading-none tracking-tighter sm:text-[40vw] md:text-[33vw] lg:text-[25vw] dark:text-white [&>span>span>span:last-child]:!hidden">
+					<SplitText text="404" delay={200} />
+				</h2>
+
+				<AnimatedContent className="relative z-50">
+					<Button
+						as={Link}
+						href="/"
+						className="max-xs:w-[90vw] only-mouse:hidden !h-14 font-medium backdrop-blur-sm md:!h-16"
+						size="lg"
+						variant="outline-touch"
+					>
+						<IconArrowLeft />
+						{__('button')}
+					</Button>
+
+					<Button
+						as={Link}
+						href="/"
+						className="max-xs:w-[90vw] only-touch:hidden !h-14 font-medium backdrop-blur-sm hover:!bg-white hover:!text-black md:!h-16"
+						size="lg"
+						variant="outline"
+					>
+						<IconArrowLeft />
+						{__('button')}
+					</Button>
+				</AnimatedContent>
+			</div>
+		</>
 	)
 }
